Revert "Move unit tests to android for coverage"
This reverts commit a36fba59731693e4d113c068ba7ab928a709da63.
