+++ data/src/test/java/com/github/leuludyha/data/api/LibraryApiTest.kt
@@ -1,11 +1,10 @@
 package com.github.leuludyha.data.api
 
-import androidx.test.platform.app.InstrumentationRegistry
 import com.github.leuludyha.data.FileReader
 import com.github.leuludyha.data.RequiringLibraryApiTest
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.runBlocking
+import kotlinx.coroutines.test.runTest
 import okhttp3.mockwebserver.MockResponse
 import org.junit.Test
 import java.net.HttpURLConnection
@@ -14,15 +13,9 @@ class LibraryApiTest: RequiringLibraryApiTest() {
 
     @OptIn(ExperimentalCoroutinesApi::class)
     @Test
-    fun searchGivesCorrectResultWithTwoDocuments() = runBlocking {
+    fun searchGivesCorrectResultWithTwoDocuments() = runTest {
 
-        val json = FileReader.readResourceFromFile(
-            InstrumentationRegistry
-            .getInstrumentation()
-            .context
-            .assets
-            .open("search_2docs.json")
-        )
+        val json = FileReader.readResourceFromFile(this.javaClass.classLoader!!, "search_2docs.json")
 
         val expectedResponse = MockResponse()
             .setResponseCode(HttpURLConnection.HTTP_OK)
@@ -56,15 +49,9 @@ class LibraryApiTest: RequiringLibraryApiTest() {
 
     @OptIn(ExperimentalCoroutinesApi::class)
     @Test
-    fun getWorkGivesCorrectResult() = runBlocking {
+    fun getWorkGivesCorrectResult() = runTest {
 
-        val json = FileReader.readResourceFromFile(
-            InstrumentationRegistry
-                .getInstrumentation()
-                .context
-                .assets
-                .open("getWork.json")
-        )
+        val json = FileReader.readResourceFromFile(this.javaClass.classLoader!!, "getWork.json")
 
         val expectedResponse = MockResponse()
             .setResponseCode(HttpURLConnection.HTTP_OK)
@@ -88,16 +75,9 @@ class LibraryApiTest: RequiringLibraryApiTest() {
 
     @OptIn(ExperimentalCoroutinesApi::class)
     @Test
-    fun getEditionGivesCorrectResult() = runBlocking {
+    fun getEditionGivesCorrectResult() = runTest {
 
-        val json = FileReader
-            .readResourceFromFile(
-                InstrumentationRegistry
-                    .getInstrumentation()
-                    .context
-                    .assets
-                    .open("getEdition.json")
-            )
+        val json = FileReader.readResourceFromFile(this.javaClass.classLoader!!, "getEdition.json")
 
         val expectedResponse = MockResponse()
             .setResponseCode(HttpURLConnection.HTTP_OK)
@@ -122,15 +102,9 @@ class LibraryApiTest: RequiringLibraryApiTest() {
 
     @OptIn(ExperimentalCoroutinesApi::class)
     @Test
-    fun getAuthorGivesCorrectResult() = runBlocking {
+    fun getAuthorGivesCorrectResult() = runTest {
 
-        val json = FileReader.readResourceFromFile(
-            InstrumentationRegistry
-                .getInstrumentation()
-                .context
-                .assets
-                .open("getAuthor.json")
-        )
+        val json = FileReader.readResourceFromFile(this.javaClass.classLoader!!, "getAuthor.json")
 
         val expectedResponse = MockResponse()
             .setResponseCode(HttpURLConnection.HTTP_OK)
